Corrected detection of Ansible library
- moves prerun ansible check as soon as possible in order to avoid
  other imports from failing before runtime checks are verified.
- adds tests that verifies special exit code is returned when Ansible
  is not installed
