+++ tox.ini
@@ -113,3 +113,5 @@ commands =
   twine check {toxinidir}/dist/*
   # Install the wheel
   sh -c "python3 -m pip install {toxinidir}/dist/*.whl"
+  # Check if ansible-lint fails due to missing ansible
+  python3 tools/check-missing-ansible.py
